+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Inclusion and Exclusion of teams within a recruitment page.
 ### Changed
 - Term of office is now visible in 'my applications' 
 ### Fixed